third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/flex-aspect-ratio-img-row-013-expected.txt

This is a testharness.js-based test.
PASS img 1
FAIL img 2 assert_equals: 
<img src="support/200x200-green.png" style="margin-bottom: 20px" data-expected-height="100" data-expected-width="100">
width expected 100 but got 120
PASS img 3
Harness: the test ran to completion.
